Remove header files dependency
