+++ src/runtime_src/core/include/CMakeLists.txt
@@ -15,7 +15,8 @@ set(XRT_HEADER_SRC
   xclhal2_mem.h
   xrt_mem.h
   xrt_error_code.h
-  xgq.h)
+  xgq.h
+  xgq_impl.h)
 
 install (FILES ${XRT_HEADER_SRC} DESTINATION ${XRT_INSTALL_INCLUDE_DIR} COMPONENT ${XRT_DEV_COMPONENT})
